requirements.txt
python-telegram-bot[job_queue]==20.0
requests==2.25.1